Make blasr compilable using gcc 4.8. bug 25737
[git-p4: depot-paths = "//depot/software/smrtanalysis/bioinformatics/lib/cpp/": change = 140502]

+++ alignment/suffixarray/SuffixArray.hpp
@@ -237,8 +237,8 @@ public:
             }
             endPosTable[curPrefix.tuple] = indexPos;
         }
-        while (indexPos < targetLength - lookupPrefixLength + 1 and 
-                curPrefix.tuple < lookupTableLength-1);
+        while ((indexPos < targetLength - lookupPrefixLength + 1) and 
+               (uint32_t(curPrefix.tuple) < uint32_t(lookupTableLength - 1)));
     }
 
     void AllocateSuffixArray(SAIndexLength stringLength) {
